Test Catalog Management shell surfaces: Lab Units v2.0, Panel v2.2, Test Catalog Completion v2
Registers the sibling admin surfaces of the Test Catalog Management shell (peers of Tests / Sample Types):
- Lab Units Management v2.0 (OGC-189): rebased on TEST_SECTION; no code field; Description required; domain a declared dependency (unbuilt on develop despite OGC-361). Supersedes the 2026-03 lab-units stub (in place, slug unchanged).
- Panel Management Domain Upgrade v2.2 (OGC-224): Panels context in the shell; single required domain (Clinical at launch); no code/lab-unit; sample types derived from member tests. Supersedes the 2026-03 panel stub (in place, slug unchanged).
- Test Catalog Completion v2 (OGC-1112, FR-46-86): consolidated single-handoff FRS + mockup; Part A delivered-gap corrections + Part B Manageability (grouped assay families, create-in-place, specimen variants). New entry; marks test-catalog-editor-completion superseded_by.

Superseded v1 stubs archived to designs/_archive/2026-07-15/. Tests 252/252, build green.
dist/ intentionally not committed (deploy-gallery.yml rebuilds from source; avoids conflict with the open sample-type PR).
