+++ mockup-viewer/src/App.jsx
@@ -34,6 +34,18 @@ export const MOCKUP_REGISTRY = [
     tags: ['test-catalog', 'loinc', 'ranges', 'result-types', 'i18n', 'admin', 'configuration'],
   },
   {
+    name: 'Test Catalog Completion v2',
+    category: 'admin-config',
+    component: React.lazy(() => import('@designs/admin-config/test-catalog-completion-v2.jsx')),
+    description: 'Consolidated Test Catalog FRS (v2) — the single implementing-agent handoff. Part A corrects delivered gaps in the shipped unified editor (custom label presets filtered out of the picker, terminology mappings missing display names, LOINC integrity warnings siloed in one section, half the combined shared-settings editor shipped, activation allowed on tests that cannot capture a result). Part B adds Manageability: grouped catalog view treating specimen variants of an assay as one family (link/unlink, issues toggle), create-in-place, add-specimen-variant with completeness rail + activation gate. FR-46–86. Supersedes the test-catalog-editor-completion and test-catalog-manageability work lists.',
+    specPath: 'designs/admin-config/test-catalog-completion-v2.md',
+    added: '2026-07-15',
+    updated: '2026-07-15',
+    status: 'draft',
+    jira: ['OGC-1112'],
+    tags: ['test-catalog', 'manageability', 'specimen-variant', 'loinc', 'activation-gate', 'admin', 'configuration'],
+  },
+  {
     name: 'Report Management',
     category: 'admin-config',
     component: React.lazy(() => import('@designs/admin-config/report-management.jsx')),
@@ -55,13 +67,17 @@ export const MOCKUP_REGISTRY = [
     tags: ['lookup-tables', 'reference-data', 'admin', 'configuration'],
   },
   {
-    name: 'Lab Units',
+    name: 'Lab Units Management',
     category: 'admin-config',
     component: React.lazy(() => import('@designs/admin-config/lab-units.jsx')),
-    description: 'Laboratory units configuration',
+    description: 'Lab Units Management (OGC-189) — v2.0, rebased on the real TEST_SECTION data model and aligned to the Test Catalog Management shell as a peer of Tests / Panels / Sample Types. Terminology is "Lab Unit" (never "test section"); no code field (no column exists); Description is required (NOT NULL). Domain is a declared dependency — unbuilt on develop despite OGC-361 (no test_section migration exists), so the single Clinical/Environmental/Vector domain requires a migration (backfill existing rows to Clinical, OGC-936 pattern).',
     specPath: 'designs/admin-config/lab-units.md',
+    added: '2026-03-03',
+    updated: '2026-07-15',
+    status: 'draft',
     githubIssue: 2,
-    tags: ['units', 'measurement', 'admin', 'configuration'],
+    jira: ['OGC-189'],
+    tags: ['lab-units', 'test-section', 'test-catalog', 'domain', 'admin', 'configuration', 'OGC-189'],
   },
   {
     name: 'Methods',
@@ -106,13 +122,17 @@ export const MOCKUP_REGISTRY = [
     tags: ['barcode', 'labels', 'presets', 'admin', 'configuration', 'test-catalog', 'order-entry'],
   },
   {
-    name: 'Panel',
+    name: 'Panel Management',
     category: 'admin-config',
     component: React.lazy(() => import('@designs/admin-config/panel.jsx')),
-    description: 'Test panel configuration',
+    description: 'Panel Management — Domain Upgrade (OGC-224) — v2.2, aligned to the verified Test Catalog data model and Manageability decisions. Managed as a Panels context in the Test Catalog Management shell (peer of Tests / Sample Types / Lab Units). Adds a single required domain (unbuilt on develop — Dependency 1, Clinical at launch, backfill existing panels to Clinical). A panel has no code (LOINC is its identifier), no lab unit (panels span sections; scoped by domain instead), and its sample types are DERIVED from member tests (SAMPLETYPE_PANEL stays backend-synced, never surfaced).',
     specPath: 'designs/admin-config/panel.md',
+    added: '2026-03-03',
+    updated: '2026-07-15',
+    status: 'draft',
     githubIssue: 5,
-    tags: ['panels', 'test-groups', 'admin', 'configuration'],
+    jira: ['OGC-224'],
+    tags: ['panels', 'test-catalog', 'domain', 'test-groups', 'admin', 'configuration', 'OGC-224'],
   },
   {
     name: 'Range Editor',
